% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ampbox_data}
\alias{ampbox_data}
\title{Amplitude Boxplot Data}
\usage{
ampbox_data(warp_median, alpha = 0.05, ka = 1)
}
\arguments{
\item{warp_median}{fdawarp object from [time_warping()] of aligned data using
the median}

\item{alpha}{quantile value (default=.05, i.e., 95%)}

\item{ka}{scalar for outlier cutoff (default=1)}
}
\value{
Returns an `ampbox` object containing:

- `median_y`: median function
- `Q1`: First quartile
- `Q3`: Second quartile
- `Q1a`: First quantile based on alpha
- `Q3a`: Second quantile based on alpha
- `minn`: minimum extreme function
- `maxx`: maximum extreme function
- `outlier_index`: indexes of outlier functions
- `fmedian`: median function
}
\description{
This function constructs the amplitude boxplot.
}
\references{
Xie, W., S. Kurtek, K. Bharath, and Y. Sun  (2016). "A geometric
  approach to visualization of variability in functional data." Journal of
  the American Statistical Association in press: 1-34.
}
\keyword{alignment}
\keyword{boxplot}
\keyword{internal}
\keyword{srvf}
