% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_vel}
\alias{growth_vel}
\title{Berkeley Growth Velocity Dataset}
\format{
## `growth_vel`
A list with two components:

- `f`: A numeric matrix of shape \eqn{69 \times 93} storing a sample of size
\eqn{N = 93} of curves evaluated on a grid of size \eqn{M = 69}.
- `time`: A numeric vector of size \eqn{M = 69} storing the grid on which
the curves `f` have been evaluated.
}
\usage{
growth_vel
}
\description{
Combination of both boys and girls growth velocity from the Berkley dataset.
}
\keyword{datasets}
